Refactor how Scalibr distinguishes standalone extractors vs filesystem extractors.
Currently, the distinction is made explicitly in a number of places; users have to provide two lists of arguments and code has to deal with this duplication in special ways.

Instead, let's treat them all the same through the `extractor.Extractor` interface until the distinction actually matters in Scan(). Until then, we're mostly concerned with name resolution. Since Name() is defined on `extractor.Extractor`, that's perfectly fine. Once the distinction becomes relevant, we can use a type switch to produce two slices of extractors of their specific interfaces.

PiperOrigin-RevId: 638029929