Extract Windows patch level but using the registry instead of the `dism` utility.
The registry key used is `SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing\Packages`.

This approach has a few advantages:
- Faster and does not rely on running an underlying command;
- Does not require admin privileges;

But it has been tested only on a limited set of machines and needs to be tested at bigger scale to ensure feature parity.

This plugin only retrieves the version associated with the packages and does not retrieve the image version (the current OS version), which needs to be implemented in a separate plugin.

PiperOrigin-RevId: 640160237