detector/binary/govulncheck: report only symbol level vulns
Currently, every single OSV reported by govulncheck JSON is converted to
a scalibr finding. This seems incorrect. govulncheck JSON will emit all
OSVs that apply to module dependencies, regardless of the version at
which the binary is using them. The results of version checking
manifest as govulncheck findings at the scan module level. Further, we
are interested in govulncheck findings at the scan symbol level. This CL
takes govulncheck findings at the symbol level (i.e., detected symbols
in the binary) and converts those to scalibr findings.

We also add tests that fail for the code before this change, but succeed
with this change.