Move inventories + findings under a top-level struct inside ScanResults.
This allows us to add new result types for things that are not software packages or security findings (e.g. running processes, open ports).

In a follow-up change we'll make extractor plugins return generic inventory types instead of just software packages.

Note that we still use the software package struct to store non-software inventory such as containerd runtimes. Moving that to a separate type will be done in a follow-up change.

PiperOrigin-RevId: 733262169