feat(guided remediation): Add relax strategy to FixVulns
Added the necessary bits to allow npm `package.json` files to be remediated with the relax strategy.

Right now it's missing the ability to write `package-lock.json` (by shelling out to `npm`) that osv-scanner has.

As part of this, made a combined native dependency client to allow FixVulns to lazily load the relevant native dependency client (maven, npm, pypi) to use so you don't need to know it beforehand.

PiperOrigin-RevId: 747641924