feat(guided remediation): Add PURLs to JSON output.
Adds `purl` to package field, and `purlFrom` & `purlTo` to package updates fields in the JSON output.

These strings are manually populated when constructing the structs, and I don't have any checks that make sure the packages match the PURLs.

I've also done a small refactor to be able to reuse PURL conversion code in guided remediation.

PiperOrigin-RevId: 750398863