feat(guided remediation): Regenerate `package-lock.json` if provided for relax strat
Shell out to `npm install` (if it's installed) if both `package.json` and `package-lock.json` files are provided to `FixVulns`.

This shelling out currently lives in `guidedremediation.go`, but this may need to move in the future if we want to support this with the npm upgrader (or if we manually implement the `package-lock.json` regeneration)

Also, if both are provided, compute the vulnerabilities that are fixed by the relock alone for the relax strategy. These are listed in the json patches as a patch with no `packageUpgrades`.

PiperOrigin-RevId: 756748518