feat(guided remediation): `FixVulnsInteractive`
Ports the interactive guided remediation TUI from OSV-Scanner to `FixVulnsInteractive` so that we can completely migrate the guided remediation logic code out.

I've just done some quick manual testing to verify it still works, but it's hard to tell if there were any changes in behaviour. I don't think it really matters if the interactive mode changes a bit anyway.

Sorry the size of this is absolutely colossal. The things in the `internal/tui` directory are mostly copied over from OSV-Scanner, though there are some minor changes to make it work with other changes I've made during the migration.

PiperOrigin-RevId: 766953254