Introduce the Veles secret scanning library and rudimentary Scalibr integration.
Veles is a standalone library for secret scanning that ships as part of Scalibr. It can detect and (where possible) validate credentials and other things colloquially referred to as "secrets".

It can be used as part of Scalibr via a corresponding Extractor and Enricher. Some parts of the integration are still under development.

PiperOrigin-RevId: 773318690