Plugin to detect end-of-life linux distributions
A proof of concept for detecting end-of-support operating systems.

Distributions that are not supported anymore pose a security risk as
they don't receive security updates. From a remediation point of view,
it often makes sense to upgrade/reinstall the OS rather than individual
packages.

Detecting end-of-support operating systems can be tricky and can't
always be done just by parsing the version ID. For example Ubuntu might
have some "extended security maintenance" enabled (Ubuntu Pro) that
extends the support for 5 additional years.

In this PoC, I implemented a single detector that detects fedora and
ubuntu releases that are not supported anymore. It's also possible to
split the detector into individual ones if needed.

Let me know if you think this is a feature that has its place in Scalibr
and I'd be happy to improve it.

The obvious disadvantage of this approach is that the plugin needs to be
updated whenever there's a new release instead of relying on an external
service or feed.

Some resources:
 - https://ubuntu.com/about/release-cycle
 - https://docs.fedoraproject.org/en-US/releases/eol/