[extractor/rpm] Change purl namespace for Amazon RPMs
Change namespace from `amzn` to `amazon` for amazonlinux RPMs.

Package-url generation is based on `/etc/os-release` `ID` field.
While this works well for fedora, centos etc, amazonlinux is a bit different, as it uses the ID `amzn`.

In this case `amzn` is more a shortname like `el` (centos) or `fc` (fedora)
While the purl specification does not expicitely list `amazon` as the correct namespace in
https://github.com/package-url/purl-spec/blob/main/PURL-TYPES.rst#rpm

It says "The namespace is the vendor such as Fedora or OpenSUSE. It is not case sensitive and must be lowercased.".
So it's probably better to use `amazon` rather than `amzn`.

Also most (all?) SBOM generation tools are using `amazon` (ex.: trivy or vulncheck).