fix: Local image scanning by calling cleanup
Turns out runtime.AddCleanup is sneaky. Because the last instance of `img` being used is on line scalibr.go:288 , go optimises it away in the following lines leaving it free for the GC to remove it and therefore call the cleanup function early (sometimes), while we are still using the layers.

By calling cleanup correctly at the end, it keeps the pointer around so the GC cleanup will not be called early.

PiperOrigin-RevId: 780784160