Fix array overindexing bug in diriterate when using the fallback method on empty dirs.
Found in https://github.com/google/osv-scalibr/pull/1291 - if someone implements their own FS that doesn't support fs.ReadDirFile we fall back to reading the files into an array. However, the interface might return "nil" to indicate there are no files in this dir.
Previously we interpreted i.files==nil as the fs.ReadDirFile existing. However, this could also mean that there are no files in the directory.
This change adds another check for the fs.ReadDirFile interface not existing, in which case we return EOF immediately when trying to iterate through the empty dir.

PiperOrigin-RevId: 805319752