Only run the COS duplicate annotator on COS filesystems.
The annotator now checks for the presence of `/etc/cos-package-info.json`. If this file does not exist, the annotator will not perform any annotations, as it indicates the filesystem is not a COS image.

The current implementation marks all packages as "ExploitabilitySignals{MatchesAllVulns: true}", even when run on non-cos OSes.

PiperOrigin-RevId: 840367129